compact id label

+++ include/hnswlib/hnswlib.h
@@ -122,7 +122,7 @@ static bool AVX512Capable() {
 #include <string.h>
 
 namespace hnswlib {
-typedef size_t labeltype;
+typedef uint64_t labeltype;
 
 // This can be extended to store state for filtering (e.g. from a std::set)
 class BaseFilterFunctor {
